refactor: add more exceptions handlers for books.

+++ tests/test_books/test_delete_books.py
@@ -61,15 +61,13 @@ def test_dont_delete_book_with_active_loan(client, token, book_db, loan):
     )
 
     assert response.status_code == HTTPStatus.CONFLICT
-    assert response.json() == {
-        'detail': 'Book is currently on loan. Cannot delete him.'
-    }
+    assert response.json() == (
+        f'Book (ISBN [{book_db.isbn}]) is currently on loan. Cannot delete him.'
+    )
 
 
 def test_delete_book_not_found(client, token):
-    response = client.delete(
-        '/books/-1', headers={'Authorization': f'Bearer {token}'}
-    )
+    response = client.delete('/books/-1', headers={'Authorization': f'Bearer {token}'})
 
     assert response.status_code == HTTPStatus.NOT_FOUND
-    assert response.json() == f'Book with ISBN -1 not found. Verify.'
+    assert response.json() == 'Book (ISBN [-1]) not found. Verify.'
